+++ yeat/cli/cli.py
@@ -93,21 +93,20 @@ def workflow_configuration(parser):
     )
 
 
-CONFIG_TEMPLATE = {
-    "global_settings": {
-        "coverage_depth": 150,
-        "downsample": -1,
-        "genome_size": 0,
-        "min_length": 100,
-        "quality": 10,
-        "skip_filter": False,
-    },
-    "samples": {"sample1": {"illumina": "short_reads_?.fastq.gz"}},
-    "assemblers": {"spades_default": {"algorithm": "spades"}},
-}
-
-
 class InitAction(Action):
+    config_template = {
+        "global_settings": {
+            "coverage_depth": 150,
+            "downsample": -1,
+            "genome_size": 0,
+            "min_length": 100,
+            "quality": 10,
+            "skip_filter": False,
+        },
+        "samples": {"sample1": {"illumina": "short_reads_?.fastq.gz"}},
+        "assemblers": {"spades_default": {"algorithm": "spades"}},
+    }
+
     def __call__(self, parser, namespace, values, option_string=None):
-        toml.dump(CONFIG_TEMPLATE, sys.stdout)
+        toml.dump(self.config_template, sys.stdout)
         raise SystemExit()