+++ yeat/tests/test_config.py
@@ -7,12 +7,28 @@
 # Development Center.
 # -------------------------------------------------------------------------------------------------
 
+from pathlib import Path
+from pydantic import ValidationError
 import pytest
 from yeat.config.config import ConfigurationError, AssemblyConfiguration
 from yeat.config.global_settings import GlobalSettings
+from yeat.config.sample import Sample
+
+
+def test_has_one_sample():
+    message = "Config has no samples"
+    with pytest.raises(ValidationError, match=message):
+        AssemblyConfiguration(global_settings=GlobalSettings, samples={}, assemblers={})
+
+
+def test_has_one_assembler():
+    samples = {"sample1": Sample(label="sample1", data={"ont_simplex": [Path("DNE")]})}
+    message = "Config has no assemblers"
+    with pytest.raises(ValidationError, match=message):
+        AssemblyConfiguration(global_settings=GlobalSettings, samples=samples, assemblers={})
 
 
 def test_algorithm_not_supported():
-    message = "unknown assembly algorithm algorithm_DNE"
+    message = "Unknown assembly algorithm DNE"
     with pytest.raises(ConfigurationError, match=message):
-        AssemblyConfiguration.select("algorithm_DNE")
+        AssemblyConfiguration.select("DNE")