+++ yeat/tests/test_sample.py
@@ -8,19 +8,41 @@
 # -------------------------------------------------------------------------------------------------
 
 from shutil import copy
+from pathlib import Path
+from pydantic import ValidationError
 import pytest
 from yeat.config.sample import SampleConfigurationError, Sample
 from yeat.tests import data_file
 
 
-def test_expand_read_paths_unable_to_find():
+@pytest.mark.parametrize("data", [{}, {"downsample": -1}])
+def test_has_one_read_type(data):
+    message = "Sample must have at least one read type"
+    with pytest.raises(ValidationError, match=message):
+        Sample(label="sample1", data=data)
+
+
+@pytest.mark.parametrize(
+    "read_path,expected",
+    [
+        ("short_reads_?.fastq.gz", ["short_reads_1.fastq.gz", "short_reads_2.fastq.gz"]),
+        ("short_reads_1.fastq.gz", ["short_reads_1.fastq.gz"]),
+    ],
+)
+def test_expand_read_path(read_path, expected):
+    data = {"illumina": data_file(read_path)}
+    Sample._expand_read_path(data)
+    assert data["illumina"] == [Path(data_file(read)) for read in expected]
+
+
+def test_expand_read_path_unable_to_find():
     data = {"illumina": "DNE"}
     message = "Unable to find fastq files"
     with pytest.raises(SampleConfigurationError, match=message):
-        Sample._expand_read_paths(data)
+        Sample._expand_read_path(data)
 
 
-def test_expand_read_paths_found_too_many(tmp_path):
+def test_expand_read_path_found_too_many(tmp_path):
     wd = tmp_path
     read1 = data_file("short_reads_1.fastq.gz")
     read2 = data_file("short_reads_2.fastq.gz")
@@ -30,7 +52,7 @@ def test_expand_read_paths_found_too_many(tmp_path):
     data = {"illumina": str(wd / "short_reads_*.fastq.gz")}
     message = "Found too many fastq files"
     with pytest.raises(SampleConfigurationError, match=message):
-        Sample._expand_read_paths(data)
+        Sample._expand_read_path(data)
 
 
 @pytest.mark.parametrize(
@@ -45,5 +67,5 @@ def test_expand_read_paths_found_too_many(tmp_path):
     ],
 )
 def test_best_long_read_type(data, read_type):
-    sample = Sample(label="testing", data=data)
+    sample = Sample(label="sample1", data=data)
     assert sample.best_long_read_type == read_type